pyproject.toml
[tool.poetry]
name = "professor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
RateMyProfessorPyAPI = "^1.2.1"
RateMyProfessorAPI = "^1.2.0"
replit = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
